Support GL_GLOB for TAG key
The TAG key can be used in rules for event matching. At the moment, it
does not support wildcards (i.e. GL_GLOB). While the key is setup with
the correct glob, the match test uses streq() which only considers
identical strings.

This patch changes the test to use match_key(), which will assess the
glob type and perform the matching accordingly.

Signed-off-by: Felipe Franciosi <felipe@paradoxo.org>